Redesigned folder structure for api project
Redesigned the folder structure of the api project to match the go standard. Created separate files for each handler and service function.
