--- api/controllers/endpoints.go
@@ -1,23 +1,0 @@
-package controllers
-
-import (
-	"myanimevault/middleware"
-
-	"github.com/gin-gonic/gin"
-)
-
-func RegisterEndpoints(server *gin.Engine) {
-	//auth routes
-	server.GET("/api/users/getCurrentUser", middleware.Authenticate, getCurrentUser)
-	server.POST("/api/users/register", register)
-	server.POST("/api/users/login", login)
-	server.GET("/api/users/refresh", refresh)
-	server.DELETE("/api/users/logout", middleware.Authenticate, logout)
-
-	//userAnime routes
-	server.GET("/api/user/anime", middleware.Authenticate, GetUserAnimeList)
-	server.GET("/api/user/anime/:animeId", middleware.Authenticate, GetUserAnimeDetails)
-	server.POST("/api/user/anime", middleware.Authenticate, CreateUserAnime)
-	server.PATCH("/api/user/anime/:animeId", middleware.Authenticate, UpdateUserAnime)
-	server.DELETE("/api/user/anime/:animeId", middleware.Authenticate, DeleteUserAnime)
-}
