+++ api/internal/handlers/useranimehandler/get_list.go
@@ -1,0 +1,21 @@
+package useranimehandler
+
+import (
+	"myanimevault/internal/services"
+	"net/http"
+
+	"github.com/gin-gonic/gin"
+)
+
+func GetUserListHandler(context *gin.Context) {
+	userId := context.GetString("userId")
+
+	animeList, err := services.GetList(userId)
+
+	if err != nil {
+		context.JSON(http.StatusInternalServerError, gin.H{"message": "There was a problem retrieving the users list"})
+		return
+	}
+
+	context.JSON(http.StatusOK, gin.H{"message": "List successfully retrieved.", "animeList": animeList})
+}
