+++ api/internal/models/dtos/coverImageDto.go
