+++ api/internal/models/dtos/userAnimeDetails.go
