+++ api/internal/models/entities/userAnime.go
