+++ api/internal/models/requests/userAnimePatchRequest.go
