+++ api/internal/services/authTokenService.go
@@ -4,9 +4,9 @@ import (
 	"database/sql"
 	"errors"
 	"fmt"
-	"myanimevault/database"
-	"myanimevault/models/customErrors"
-	"myanimevault/utils"
+	"myanimevault/internal/database"
+	"myanimevault/internal/models/customErrors"
+	"myanimevault/internal/utils"
 	"os"
 	"strings"
 	"time"
