+++ api/internal/utils/hash.go
